Added support for leaf nodes
Leaf nodes are displayed differently in netbeans than nodes with
children. They are marked as such by using 'Children.LEAF' as
their children object. So far, this was not supported by 'NbNode's,
since their children's type was always 'NbChildren'.

This is changed in the present commit. There are now factory methods
in 'NbChildren' to create true child nodes.