Changed logging for NbSystem
This is a temporary change only. However, it was a bad idea
to use TcPreferences for this, since one might to allow the user
to set these preferences upon startup of an application using
a dialog. Such a dialog could only be run with the reactive
system already running, which of course would lead to a cyclic
dependancy and in the end a stackoverflow.

Later this should be changed, so that the log level can still
be set by the user (through a system variable for instance).